+++ lib/stats/charts/attempts_by_date.dart
@@ -5,7 +5,9 @@ import 'package:sendrax/models/attempt.dart';
 import 'package:sendrax/util/constants.dart';
 
 class AttemptsByDateChart extends StatefulWidget {
-  AttemptsByDateChart({Key key, @required this.attempts}) : super(key: key);
+  AttemptsByDateChart({Key key, @required this.attempts, @required this.locationNamesToIds})
+      : super(key: key);
+  final Map<String, String> locationNamesToIds;
   final List<Attempt> attempts;
 
   @override
@@ -14,6 +16,10 @@ class AttemptsByDateChart extends StatefulWidget {
 
 class _AttemptsByDateChartState extends State<AttemptsByDateChart> {
   List<charts.Series> chartData;
+  String filterTimeframe;
+  String filterLocation;
+  String filterSendType;
+  String filterCategory;
 
   DateTime selectedDate;
   int selectedCount;
@@ -27,6 +33,7 @@ class _AttemptsByDateChartState extends State<AttemptsByDateChart> {
   @override
   Widget build(BuildContext context) {
     List<Widget> children = <Widget>[];
+    children.add(_buildFilters(context));
     children.add(_buildChart(context));
     if (selectedDate != null) {
       children.add(Padding(
@@ -47,52 +54,230 @@ class _AttemptsByDateChartState extends State<AttemptsByDateChart> {
         padding: EdgeInsets.all(UIConstants.SMALLER_PADDING), child: Column(children: children));
   }
 
-  Widget _buildChart(BuildContext context) {
+  Widget _buildFilters(BuildContext context) {
+    return Container(
+        child: Row(children: <Widget>[
+      Expanded(
+          child: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: <Widget>[
+          Row(
+            children: <Widget>[
+              _showTimeFrameDropdown(context),
+              _showLocationDropdown(context),
+            ],
+          ),
+          Row(
+            children: <Widget>[
+              _showSendTypeDropdown(context),
+              _showCategoryDropdown(context),
+            ],
+          )
+        ],
+      )),
+      _showClearDropdownsButton(context),
+    ]));
+  }
+
+  Widget _showTimeFrameDropdown(BuildContext context) {
     return Expanded(
-        child: Container(
-            padding: EdgeInsets.all(UIConstants.SMALLER_PADDING),
-            decoration: BoxDecoration(
-              color: Theme.of(context).cardColor,
-              borderRadius: BorderRadius.all(Radius.circular(UIConstants.CARD_BORDER_RADIUS))
-            ),
-            child: charts.TimeSeriesChart(
-              chartData,
-              dateTimeFactory: charts.LocalDateTimeFactory(),
-              selectionModels: [
-                charts.SelectionModelConfig(
-                  type: charts.SelectionModelType.info,
-                  changedListener: _onSelectionChanged,
-                )
-              ],
-              primaryMeasureAxis: charts.NumericAxisSpec(
-                  renderSpec: charts.GridlineRendererSpec(
-                labelStyle: charts.TextStyleSpec(
-                  fontSize: Theme.of(context).accentTextTheme.caption.fontSize.toInt(),
-                  fontWeight: Theme.of(context).accentTextTheme.caption.fontWeight.toString(),
-                  color: charts.ColorUtil.fromDartColor(Theme.of(context).accentColor),
-                ),
-                lineStyle: charts.LineStyleSpec(
-                  color: charts.ColorUtil.fromDartColor(Theme.of(context).dialogBackgroundColor),
-                ),
-              )),
-              domainAxis: charts.DateTimeAxisSpec(
-                  renderSpec: charts.SmallTickRendererSpec(
-                labelStyle: charts.TextStyleSpec(
-                  fontSize: Theme.of(context).accentTextTheme.caption.fontSize.toInt(),
-                  fontWeight: Theme.of(context).accentTextTheme.caption.fontWeight.toString(),
-                  color: charts.ColorUtil.fromDartColor(Theme.of(context).accentColor),
-                ),
-              )),
-            )));
+        child: Padding(
+            padding: EdgeInsets.fromLTRB(
+                0.0, 0.0, UIConstants.SMALLER_PADDING / 2, UIConstants.SMALLER_PADDING / 2),
+            child: Container(
+                padding: EdgeInsets.all(UIConstants.STANDARD_PADDING),
+                decoration: BoxDecoration(
+                    color: Theme.of(context).cardColor,
+                    borderRadius:
+                        BorderRadius.all(Radius.circular(UIConstants.BUTTON_BORDER_RADIUS))),
+                child: DropdownButtonHideUnderline(
+                    child: DropdownButton<String>(
+                  style: Theme.of(context).accentTextTheme.subtitle2,
+                  items: _createDropdownItems(TimeFrames.TIME_FRAMES.values.toList()),
+                  value: filterTimeframe,
+                  hint: Text("Time frame"),
+                  isExpanded: true,
+                  isDense: true,
+                  onChanged: (value) => setState(() {
+                    filterTimeframe = value;
+                    chartData = _buildChartData(context);
+                  }),
+                )))));
+  }
+
+  Widget _showLocationDropdown(BuildContext context) {
+    return Expanded(
+        child: Padding(
+            padding: EdgeInsets.fromLTRB(
+                UIConstants.SMALLER_PADDING / 2, 0.0, 0.0, UIConstants.SMALLER_PADDING / 2),
+            child: Container(
+                padding: EdgeInsets.all(UIConstants.STANDARD_PADDING),
+                decoration: BoxDecoration(
+                    color: Theme.of(context).cardColor,
+                    borderRadius:
+                        BorderRadius.all(Radius.circular(UIConstants.BUTTON_BORDER_RADIUS))),
+                child: DropdownButtonHideUnderline(
+                    child: DropdownButton<String>(
+                  style: Theme.of(context).accentTextTheme.subtitle2,
+                  items: _createDropdownItems(widget.locationNamesToIds.keys.toList()),
+                  value: filterLocation,
+                  hint: Text("Location"),
+                  isExpanded: true,
+                  isDense: true,
+                  onChanged: (value) => setState(() {
+                    filterLocation = value;
+                    chartData = _buildChartData(context);
+                  }),
+                )))));
+  }
+
+  Widget _showSendTypeDropdown(BuildContext context) {
+    return Expanded(
+        child: Padding(
+            padding: EdgeInsets.fromLTRB(
+                0.0, UIConstants.SMALLER_PADDING / 2, UIConstants.SMALLER_PADDING / 2, 0.0),
+            child: Container(
+                padding: EdgeInsets.all(UIConstants.STANDARD_PADDING),
+                decoration: BoxDecoration(
+                    color: Theme.of(context).cardColor,
+                    borderRadius:
+                        BorderRadius.all(Radius.circular(UIConstants.BUTTON_BORDER_RADIUS))),
+                child: DropdownButtonHideUnderline(
+                    child: DropdownButton<String>(
+                  style: Theme.of(context).accentTextTheme.subtitle2,
+                  items: _createDropdownItems(SendTypes.SEND_TYPES),
+                  value: filterSendType,
+                  hint: Text("Send type"),
+                  isExpanded: true,
+                  isDense: true,
+                  onChanged: (value) => setState(() {
+                    filterSendType = value;
+                    chartData = _buildChartData(context);
+                  }),
+                )))));
+  }
+
+  Widget _showCategoryDropdown(BuildContext context) {
+    return Expanded(
+        child: Padding(
+            padding: EdgeInsets.fromLTRB(
+                UIConstants.SMALLER_PADDING / 2, UIConstants.SMALLER_PADDING / 2, 0.0, 0.0),
+            child: Container(
+                padding: EdgeInsets.all(UIConstants.STANDARD_PADDING),
+                decoration: BoxDecoration(
+                    color: Theme.of(context).cardColor,
+                    borderRadius:
+                        BorderRadius.all(Radius.circular(UIConstants.BUTTON_BORDER_RADIUS))),
+                child: DropdownButtonHideUnderline(
+                    child: DropdownButton<String>(
+                  style: Theme.of(context).accentTextTheme.subtitle2,
+                  items: _createDropdownItems(ClimbCategories.CATEGORIES),
+                  value: filterCategory,
+                  hint: Text("Category"),
+                  isExpanded: true,
+                  isDense: true,
+                  onChanged: (value) => setState(() {
+                    filterCategory = value;
+                    chartData = _buildChartData(context);
+                  }),
+                )))));
+  }
+
+  Widget _showClearDropdownsButton(BuildContext context) {
+    return Container(
+        child: IconButton(
+            icon: Icon(Icons.cancel,
+                color: (filterTimeframe == null &&
+                        filterLocation == null &&
+                        filterSendType == null &&
+                        filterCategory == null)
+                    ? Colors.grey
+                    : Theme.of(context).accentColor),
+            onPressed: () => setState(() {
+                  filterTimeframe = null;
+                  filterLocation = null;
+                  filterSendType = null;
+                  filterCategory = null;
+                  chartData = _buildChartData(context);
+                })));
+  }
+
+  List<DropdownMenuItem> _createDropdownItems(List<String> items) {
+    if (items.isNotEmpty) {
+      return items.map((String value) {
+        return DropdownMenuItem<String>(
+          value: value,
+          child: Text(value),
+        );
+      }).toList();
+    } else {
+      // null disables the dropdown
+      return null;
+    }
+  }
+
+  Widget _buildChart(BuildContext context) {
+    Widget chart = (chartData != null)
+        ? charts.TimeSeriesChart(
+            chartData,
+            dateTimeFactory: charts.LocalDateTimeFactory(),
+            selectionModels: [
+              charts.SelectionModelConfig(
+                type: charts.SelectionModelType.info,
+                changedListener: _onSelectionChanged,
+              )
+            ],
+            primaryMeasureAxis: charts.NumericAxisSpec(
+                renderSpec: charts.GridlineRendererSpec(
+              labelStyle: charts.TextStyleSpec(
+                fontSize: Theme.of(context).accentTextTheme.caption.fontSize.toInt(),
+                fontWeight: Theme.of(context).accentTextTheme.caption.fontWeight.toString(),
+                color: charts.ColorUtil.fromDartColor(Theme.of(context).accentColor),
+              ),
+              lineStyle: charts.LineStyleSpec(
+                color: charts.ColorUtil.fromDartColor(Theme.of(context).dialogBackgroundColor),
+              ),
+            )),
+            domainAxis: charts.DateTimeAxisSpec(
+                renderSpec: charts.SmallTickRendererSpec(
+              labelStyle: charts.TextStyleSpec(
+                fontSize: Theme.of(context).accentTextTheme.caption.fontSize.toInt(),
+                fontWeight: Theme.of(context).accentTextTheme.caption.fontWeight.toString(),
+                color: charts.ColorUtil.fromDartColor(Theme.of(context).accentColor),
+              ),
+            )),
+          )
+        : Center(
+            child: Text(
+            "There are no existing attempts ${widget.attempts.isNotEmpty ? "matching these filters" : ""}. \nGo log some!",
+            style: Theme.of(context).accentTextTheme.subtitle2,
+            textAlign: TextAlign.center,
+          ));
+    return Expanded(
+        child: Padding(
+            padding: EdgeInsets.only(top: UIConstants.SMALLER_PADDING),
+            child: Container(
+                padding: EdgeInsets.all(UIConstants.SMALLER_PADDING),
+                decoration: BoxDecoration(
+                    color: Theme.of(context).cardColor,
+                    borderRadius:
+                        BorderRadius.all(Radius.circular(UIConstants.CARD_BORDER_RADIUS))),
+                child: chart)));
   }
 
   List<charts.Series<AttemptsByDateSeries, DateTime>> _buildChartData(BuildContext context) {
+    List<Attempt> filteredAttempts = _filterAttempts();
+
+    if (filteredAttempts.isEmpty) {
+      return null;
+    }
+
     List<AttemptsByDateSeries> chartData = <AttemptsByDateSeries>[];
-    DateTime firstDate = widget.attempts[0].timestamp.toDate();
+    DateTime firstDate = filteredAttempts[0].timestamp.toDate();
     DateTime currentDate = DateTime(firstDate.year, firstDate.month, firstDate.day);
     int currentCount = 1;
 
-    for (Attempt attempt in widget.attempts.skip(1)) {
+    for (Attempt attempt in filteredAttempts.skip(1)) {
       DateTime attemptDate = attempt.timestamp.toDate();
       DateTime startOfDay = DateTime(attemptDate.year, attemptDate.month, attemptDate.day);
       if (currentDate != startOfDay) {
@@ -114,6 +299,45 @@ class _AttemptsByDateChartState extends State<AttemptsByDateChart> {
         data: chartData,
       )
     ];
+  }
+
+  List<Attempt> _filterAttempts() {
+    List<Attempt> filteredAttempts = widget.attempts;
+
+    if (filterTimeframe != null) {
+      if (filterTimeframe == TimeFrames.TIME_FRAMES["lastWeek"]) {
+        filteredAttempts = filteredAttempts
+            .where((attempt) =>
+                DateTime.now().difference(attempt.timestamp.toDate()) < Duration(days: 7))
+            .toList();
+      } else if (filterTimeframe == TimeFrames.TIME_FRAMES["lastMonth"]) {
+        filteredAttempts = filteredAttempts
+            .where((attempt) =>
+                DateTime.now().difference(attempt.timestamp.toDate()) < Duration(days: 30))
+            .toList();
+      } else if (filterTimeframe == TimeFrames.TIME_FRAMES["lastYear"]) {
+        filteredAttempts = filteredAttempts
+            .where((attempt) =>
+                DateTime.now().difference(attempt.timestamp.toDate()) < Duration(days: 365))
+            .toList();
+      }
+    }
+    if (filterLocation != null) {
+      filteredAttempts = filteredAttempts
+          .where((attempt) => attempt.locationId == widget.locationNamesToIds[filterLocation])
+          .toList();
+    }
+    if (filterSendType != null) {
+      filteredAttempts =
+          filteredAttempts.where((attempt) => attempt.sendType == filterSendType).toList();
+    }
+    if (filterCategory != null) {
+      filteredAttempts = filteredAttempts
+          .where((attempt) => attempt.climbCategories.contains(filterCategory))
+          .toList();
+    }
+
+    return filteredAttempts;
   }
 
   _onSelectionChanged(charts.SelectionModel model) {